requirements.txt
flask
gunicorn
google-cloud-storage
google-cloud-firestore
Wand
torch
torchvision
torchaudio
ultralytics
numpy
pandas
paddleocr
